+++ packages/gatsby/src/redux/persist.ts
@@ -1,4 +1,5 @@
 import path from "path"
+import os from "os"
 import v8 from "v8"
 import {
   existsSync,
@@ -8,7 +9,7 @@ import {
   renameSync,
   writeFileSync,
 } from "fs-extra"
-import { IReduxNode, ICachedReduxState } from "./types"
+import { IGatsbyNode, ICachedReduxState } from "./types"
 import { sync as globSync } from "glob"
 import report from "gatsby-cli/lib/reporter"
 
@@ -54,7 +55,7 @@ export function readFromCache(): ICachedReduxState {
     reduxChunkedNodesFilePrefix(reduxCacheFolder) + `*`
   ).map(file => v8.deserialize(readFileSync(file)))
 
-  const nodes: [string, IReduxNode][] = [].concat(...chunks)
+  const nodes: [string, IGatsbyNode][] = [].concat(...chunks)
 
   if (!chunks.length && process.env.GATSBY_DB_NODES !== `loki`) {
     report.info(
@@ -69,7 +70,7 @@ export function readFromCache(): ICachedReduxState {
   return obj
 }
 
-function guessSafeChunkSize(values: [string, IReduxNode][]): number {
+function guessSafeChunkSize(values: [string, IGatsbyNode][]): number {
   // Pick a few random elements and measure their size then pick a chunk size
   // ceiling based on the worst case. Each test takes time so there's trade-off.
   // This attempts to prevent small sites with very large pages from OOMing.
@@ -106,7 +107,7 @@ function prepareCacheFolder(
 
   if (map) {
     // Now store the nodes separately, chunk size determined by a heuristic
-    const values: [string, IReduxNode][] = [...map.entries()]
+    const values: [string, IGatsbyNode][] = [...map.entries()]
     const chunkSize = guessSafeChunkSize(values)
     const chunks = Math.ceil(values.length / chunkSize)
 
@@ -139,7 +140,7 @@ export function writeToCache(contents: ICachedReduxState): void {
   // Note: this should be a transactional operation. So work in a tmp dir and
   // make sure the cache cannot be left in a corruptable state due to errors.
 
-  const tmpDir = mkdtempSync(`reduxcache`) // linux / windows
+  const tmpDir = mkdtempSync(path.join(os.tmpdir(), `reduxcache`)) // linux / windows
 
   prepareCacheFolder(tmpDir, contents)