+++ www/src/data/tags-docs.js
@@ -35,7 +35,7 @@ export const TAGS_AND_DOCS = new Map([
   [`documentation`, `/docs/`],
   [`documentation-sites`, ``],
   [`drupal`, `/docs/sourcing-from-drupal/`],
-  [`ecommerce`, ``],
+  [`e-commerce`, ``],
   [`enterprise`, `/docs/gatsby-in-the-enterprise/`],
   [`gatsby-apis`, `/docs/api-reference/`],
   [`gatsby-cloud`, ``],
@@ -94,4 +94,5 @@ export const TAGS_AND_DOCS = new Map([
   ],
   [`wordpress`, `/docs/sourcing-from-wordpress/`],
   [`100-Days-of-Gatsby`, ``],
+  [`lambda`, ``],
 ])